requirements.txt
beautifulsoup4>=4.12
python-dotenv>=1.0
requests>=2.31
supabase>=2.4
yfinance>=0.2
